Change imports after renaming of main file to plugin (external plugins).
