+++ spyder/plugins/breakpoints/__init__.py
@@ -3,4 +3,4 @@
 #==============================================================================
 # The following statement is required to register this 3rd party plugin:
 #==============================================================================
-from .breakpoints import Breakpoints as PLUGIN_CLASS
+from .plugin import Breakpoints as PLUGIN_CLASS
