+++ spyder/plugins/io_hdf5/__init__.py
@@ -1,7 +1,7 @@
 # =============================================================================
 # The following statements are required to register this I/O plugin:
 # =============================================================================
-from .hdf5 import load_hdf5, save_hdf5
+from .plugin import load_hdf5, save_hdf5
 
 FORMAT_NAME = "HDF5"
 FORMAT_EXT  = ".h5"
