+++ spyder/plugins/profiler/__init__.py
@@ -3,4 +3,4 @@
 #==============================================================================
 # The following statement is required to register this 3rd party plugin:
 #==============================================================================
-from .profiler import Profiler as PLUGIN_CLASS
+from .plugin import Profiler as PLUGIN_CLASS
