+++ spyder/plugins/pylint/__init__.py
@@ -3,4 +3,4 @@
 #==============================================================================
 # The following statement is required to register this 3rd party plugin:
 #==============================================================================
-from .pylint import Pylint as PLUGIN_CLASS
+from .plugin import Pylint as PLUGIN_CLASS
